requirements.txt
torch
torchvision
Pillow
fastapi
uvicorn
python-multipart
faiss-cpu
redis